Some small improvements to cobaltb.py and the readme.
(1) cobaltb.py: Be more flexible about the contents of personal_cluster.json.
It is now allowed to have missing entries (which are treated as having
the empty string for a value) and spurious entries (which are ignored.)

(2) cobaltb.py: Replace the argument --bigtable_project_name with the pair
of arguments --cloud_project_prefix and --cloud_project_name. This allows
us to eliminate the redundant field bigtable_project_name from
personal_cluster.json since that file already contains
cloud_project_prefix and cloud_project_name.

(3) cobaltb.py: The readme file describes that it is possible to run the
demo manually with the processes running locally but connecting to
Cloud Bigtable. This was actually not true because cobaltb.py did not
accept the appropraite arguments for the run command. Now it does.

(4) README.md
  (a) Change documentation to reflect the above changes
  (b) Fix some typos

Change-Id: Ied4f582b6787664c6c5dffdbcc446d38b900c3e3