Adds RapporCandidateList to report_configs.proto.
This is in support of String RAPPOR. In order to run a String RAPPOR
report the analyzer needs a list of candidates. We add structures
to report_configs.proto to support this.

We also eliminate the field |candidate_map_location|. In a follow-up
CL we will be building the candidate map in memory and not
caching it. We can add support for caching the candidate map
later. For now we have no need for the |candidate_map_location|
field.

Change-Id: I051d32c754ede5db12bfacd4bede9816d263b3cd