Adds functionality to build a candidate map.
This continues the work of implementing the parts of String RAPPOR
that occur prior to the use of LASSO. In terms of the python and R
Cobalt prototype this does the work of the Python function HashCandidates.

In more detail this builds a data structure that captures, for each
candidate string and each cohort, the indices of the set bits in
the Bloom filter.

- We add the new class RapporAnalyzer. This is the class that will contain
the String RAPPOR analysis code.
- We add the function AddObservation() which just delegates to
BloomBitCounter::AddObservation().
- We ad the private function BuildCandidateMap() which contains the
logic for building the candidate map.
- We add a stub for the function Analyze(). Currently this function
does nothing but invokes BuildCandidateMap(). We also added a comment:
TODO(azani, mironov) Put LASSO analysis here.

Change-Id: I06af4f808ed8aad105fab0af2fb071bf38275086