Adds indirection for including <glog/logging.h>.
On Fuchsia applications do not use google gLog for logging.
Instead they use a custom logging solution which is
similar to gLog but has different macro names.

Our solution is to continue to use the gLog macro names
but map them to the corresponding Fuchsia macros when
building for Fuchsia.

In support of this we replace direct references to
<glog/logging.h> with references to a new
file in the root directory, logging.h.

logging.h will look for whether HAVE_GLOG is defined.
If it is then it will do
 #include <glog/logging.h>
otherwise it will #define various gLOG macros to be equal to
the equivalent Fuchsia logging macros.

For Cobalt not on Fuchsia this should have no affect.

Note that we do not attempt to make this replacement in
every file in our code base because that would be a lot
of work and is not necessary. We can do that slowly over
time if we wish. For now we are only editing the files that
will be part of the Fuchsia client.

Change-Id: Ic53506db9cb4b140a9ccba3f26320f0d006ee737