Adds the ability to manage production Cobalt clusters.
- There is a new directory called "production". This has sub-directories
corresponding to each of our prod clusters.
- Currently there is one prod cluster called fuchsia-cobalt-us-central1-b.
- Production directories contain a file named cluster.json. This serves
a purpose similar to personal_cluster.json in the root directory except
that the production cluster.json files are checked in to source control.
- cobaltb.py has been enhanced to accept the flag --production_dir.
when this is passed the cluster.json file in the specified production dir
will be read in order to obtain default flag values used in deployment.
- The production directories are used for a couple of other purposes:
= Each developer who wants to manage a production cluster can put a
service account credential file for the prod cluster in that directory
on their local machine (not checked in).
This is used to allow the developer to use Cobalt tools to
directly manage the production Bigtables.
= A developer who wants to manage a production cluster can put PEM files there
on their local machine (not checked in) as a staging area for
deployment to GKE secrets.
- We changed the name of service_account_credentials.json to
personal_bt_admin_service_account.json to better explain its purpose.
- Many of the deployment commands in cobaltb.py delegate to the
kubectl tool. This tool has the convenient but dangerous property that
it keeps state on your machine in a config file. In particular it keeps
the state of what cluster is the "current context". This allows you to
not have to specify the context when issuing kubectl commands. But this
property is very dangerous for us because it means a developer could
accidentally manipulate a production cluster when he believes he is
managing his personal development cluster. To compensate for this
issue we know explicitly pass the --context flag to all kubectl commands
overriding the current context.
- We added the ability for cobaltb.py run test_app, observation_querier
and report_client to understand the flag --production_dir and respond
by operating apporpriately against the prod cluster. This allows
manually testing of a production cluster. The second two of these
are read-only and so are useful at any time during the life of a production
cluster. The first one writes into the cluster and so is only useufl
when a new production cluster is first created and we want to test that
it is working. (We have other commands to delete the test data).
- We update README.md to explain all of this.

Change-Id: Ie25b818ec68cb50158c8eb2ff5a379ab8c64168d