Adds support for static IP addresses.
In cluster.json there are now fields for the static
IP addresses of the Shuffler, ReportMaster and Analyzer Service.
If populated these cause the processes to use these
static IP addresses when deployed. In order for this
to work the IP addresses must already have been allocated
in the Google Cloud project. If the static IP addresses
are not specfied then dynamic IP addresses will be used.

Change-Id: I430b2318a52abdf18e65ee3a4037586792763f81