Change the Ledger project to have ID 100.
We want to be able to run tests against the production instance of Cobalt.
In doing so our number 1 priority must be to not accidentally mess up
any real production data.

One way to minimize the chances of messing up real production data is to
use project IDs for real production projects that are unlikely to be used
by accident. But project ID 1 is very likely to be used by accident.

Therefore lets reserve project IDs 1 through 99 as test-only projects.
Let's have our production projects start with project ID 100.

In this CL I am initializing this scheme with the following:

Project ID 1: Used for running the end-to-end test (in the Cobalt repo)
against the production instance of Cobalt.

Project ID 2: Used for running the manual Cobalt Test App (in the fuchsia repo)
against the production instance of Cobalt.

Project ID 100: The Ledger.

We can add further testing projects as necessary.

Change-Id: I77c8efebafe2c41a0d91c6e8a9146217017f2c5d