Support running the end-to-end test against the prod cluster.
We add support in cobaltb.py for the command
cobaltb.py --production_dir=<dir> test --tests=e2e.

This causes the end-to-end test to run against the production
Cobalt cluster.

We also clean up related parts of cobaltb.py
= We change the flag -cobalt-on-gke to be -cobalt_on_personal_cluster
because that it what it is for now that we have the
--production_dir flag.

= we factor the function _add_cloud_access_args() out of the function
= _add_gke_deployment_args() because the former is used for
things other than deployment.

Change-Id: I3d1d9d55761e95416647535deffa4249f12306a1