Adds the method ObservationStore::DeleteAllForMetric().
This method deletes all of the observations in the store corresponding
to the given metric. We plan to use this from within the bigtable_tool
in place of bigtable_store->DeleteAllRows(). We want to be able to
run the end-to-end test against the production cluster. But the test
starts off by deleting everything in Bigtable. Instead we want it
to only surgically delete the data that it knows to be test data and
not production data.

Change-Id: If316ec6c6768c2f8f4976785da21428bd5c33fc2