Changes bigtable_tool to use a more selective delete.
The tool called bigtable_tool is used for two purposes:
(a) It may be invoked directly via cobaltb.py. This is a useful thing
to do when running a manual demo. It is also used initially when
provisioning a new cluster.
(b) It is invoked from the end-to-end test in order to delete data
from Bigtable prior to running the test.

Previously this tool worked by invoking
BigtableStore::DeleteAllRows(). With this CL we change the tool
to instead invoke ObservationStore::DeleteAllForMetric() and
ReportStore::DeleteAllForReportConfig().

Thus the tool is being changed to do a more selective delete.
Correspondingly four new flags have been added to the tool in
order to pass the information about which metric or which
report config to work with.

NOTE: The code enforces that the project ID must be less than 100.
If we follow the convention that all real customer projects have
IDs >= 100 this will prevent accidental deletion of customer data.

The clients of the tool, namely cobaltb.py and the end-to-end test
have also been modified to pass these flags.

Also the end-to-end test has been given a coloful, animated
warning panel to allow a human operator to interrupt it if
it was accidentally invoked with incorrect flags.

Change-Id: Idf1d01fa8103c862fe996b3d58ce797357330fee