Make project ID settable in Observation Querier.
Observation Querier is a Cobalt tool used to query the
Observation Store. It may be run against the production instance
of Cobalt via:
./cobaltb.py --production_dir=production/fuchsia-cobalt-us-central1-b start observation_querier.

This is a read-only tool so there is no danger of corruption.

This CL enables the project ID to be set interactively so that the tool may
be used for examining the Observation store of our various customer projects.

Change-Id: I8613a76c45087059c21b0e4ee7df16a8b08350c9