Adds new script upload_report_client.py
We need a mechanism for distributing the binary report_client to
our users. This CL introduces a script that uses the standard
depot_tools mechanism of uploading a binary to Google Cloud Storage
under a name that is the sha1 of the file.

We upload to the bucket fuchsia-build/cobalt/report_client/<platform>.

We check-in the generated ".sha1" file that contains the sha1
for the linux64 binary. We still need to do this on a Mac.

Change-Id: Ic188aea8d78a320c89601984dd84150ebaa92674