Introduces the notion of the INDEX data type.
Values of type INDEX are zero-based indices into some enumerated
set that is specified outside of Cobalt's configuration. Cobalt
is not interested in the values or their data type except for the
need to use a label at report-generation time. Human-readable
labels may be assigned to a subset of the indices as part of
a ReportConfig. If labels are missing it is not a fatal error.
Instead at report-generation time a label such as
"<value 42 (undedfined)>" will be used.

The INDEX datatype may be used with Basic RAPPOR. In the future
we expect to introduce additional encodings that will also
be able to use it.

The ReportConfig proto message has been refactored to allow
the specification of labels for indices. We generalized the
mechanism that we were previously using to specify a
String RAPPOR Candidate List. Now a ReportVariable includes
a map called |per_encoding_data|. This allows us to specify
data to be used at report-generation time that is specific
to a particular encoding. That mechanism may be used to
pass the candidate list for String RAPPOR and it may be used
to pass the labels for any encoding that uses the INDEX data type
including Basic RAPPOR.

This cl includes updates to the configuration infrastruture but
does not include updates to the encoder API necessary to use
the new INDEX type during encoding. That will come in the
next CL.

Change-Id: Iae3a929affeddd0112a4f085b39f534aeb22b5ed