Adds new |label| field to HistogramReportRow.
This is a continuation of the work started in
https://fuchsia-review.googlesource.com/#/c/42789/
and
https://fuchsia-review.googlesource.com/#/c/44412/.

In those two CLs we introduced the notion of INDEX data type
and added support for it on the Encoder side. In particular
we added support for it in Basic RAPPOR.

In https://fuchsia-review.googlesource.com/#/c/42789 we
augmented the ReportConfig proto message to allow the
specification of human-readable labels associated with
indices.

In this CL we add the logic that consumes that new field
of the ReportConfig. We teach HistogramAnalysisEngine
how to read the labels and write them into the |label| field
of HistorgramReportRow, newly introduced in this CL.

To shuttle the necessary information to where it needs to go,
ReportGenerator now passes a ReportVariable* to
HistogramAnalysisEngine.

Change-Id: I2ff164b6e5b91ef0b841c4a7393f6396d00cdcc0