Adds new metric, encoding config, report config for module usage tracking.
Also I took the opportunity to clean up some formatting
and comments.

Also, in these early days I am following the practice of copying
each real metric, encoding, report into the test-only section.
This allows us to test configurations that are simlar to what
our real users are using. This practice is obviously not scalable
but given we still have only two projects I would like to continue
with it for a while.

Change-Id: I59de44c2f6dd53474cdffb4c9463471a84a7567c