Simplify ReportConfig: Eliminate ReportPerEncodingData.
A ReportConfig contains
(a) An optional list of human-readable labels associated with
the indices for a metric of type INDEX.
(b) An optional list of strings to use in string RAPPOR.

Previously I had erroneously thought that these two pieces
of data were specific to an EncodingConfig and so I needed to
introduce the notion of ReportPerEncodingData.

But I realized this was a mistake. Both (a) and (b) above are
properties of a ReportVariable, independently of which encodings
were used. So I eliminated ReportPerEncodingData and
simplified ReportConfig.

Change-Id: Ia84ee08ce0ba5d50fc6b6b27ce31cc145b9d515c