Some fixes to running the end-to-end test on the prod cluster.
We have been running the end-to-end test on the production cluster
for some time now. But I realized that this has been working
somewhat by accident.

When running the end-to-end test against production we have been
using the demo config files on the client side and the production
config files on the server side. This has been working because
the relevent parts of the demo config had been copied to the
production config. We didn't even have the facility to tell the
end-to-end test to use the production config on the client side.

In this CL we:
(a) Add a parameter to the end-to-end test for which config file to
use on the client side.
(b) Use the production config file when running the end-to-end
test against production.
(c) Add comments to the end-to-end test to explain what is going on
and to explain the fact that we need to keep project (1, 1) from
the production config files in sync with the end-to-end test.
(d) Update project (1, 1) from the production config with the latest
enhancments to the end-to-end test.

Change-Id: I3acdfb162f36252b40f6caf969aa659027db33c1