Supress empty rows in printed reports.
We would like to move the Ledger team from their current string-based
encoding to the new index-based encoding. As part of this we would
like to declare the encoding config to have a large number of
category indices so that the Ledger team can easily start using
additional categories without first talking to us. They can
then update the labels in the ReportConfig at their leisure in order
to get the new category names in their reports, but the reports
will work just fine before they do that using default labels
of the form "<index 42>" for the categories that have not yet
been assigned labels.

Currently I have defined the encoding config to have 128 categories,
and currently the first 5 of the 128 categories have been assigned
labels.

One drawback of this approach is that the printed report contains
123 essentially empty rows of the form:
<category 6>,0.000
<category 7>,0.000
...
<category 127>,0.000

In this CL we rectify this situation by introducing the parameter
supressEmptyRows in the function ReportToStrings() in report_client.
When true we will use heuristics to supress rows from the printed
output which are intutively "empty" meaning they are likely to contain
no information of interest to a reader. Our first heuristic is
that rows as described above are considered empty.

In doing this we also slightly refactored report_client.go.

Change-Id: I9026399ebc2bb3395e6f7cafbf2f515b6fb6a34b