Introduces SendRetryer.
This is a new class that encapsulates the retry logic for sending
gRPCs to the Shuffler from an Encoder client.

Callers can specify the initial_rpc_deadline and an overall deadline.
The Retryer will continue to retry, using exponential backoff,
until the overall deadline. It will increase the RPC timeout
whenever DEADLINE_EXCEEDED is returned from the Shuffler.

We introduce a ClockInterface that is used to help test
the Retryer. The ClockInterface will be useful for other tests
in the future.

Change-Id: I4607ca4c2ba2515e3b6c1433f7d248e68a0a998a