Introduces the ShippingManager.
ShippingManager is a central coordinator for collecting encoded Observations
and sending them to the Shuffler. Observations are accumulated in memory
and periodically sent in batches to the Shuffler by a background worker
thread on a regular schedule. ShippingManager also performs expedited
off-schedule sends when too much unsent Observation data has accumulated.
A client may also explicitly request an expedited send.

Change-Id: I9df9b26b67b26f69ebed656b2b4689ce9a89881e