ShippingManager: Don't send the first Observation immediately.
Previously we were initializing next_scheduled_send_time_ to zero.
This meant that after a single Observation was added to the
ShippingManager it was sent immediately and then
next_scheduled_send_time_ was set to the current time plus
schedule_interval_ and so the second Observation had to wait
for a scheduled send.

Here we simplify this by initializing
next_scheduled_send_time_(std::chrono::system_clock::now() +
                                schedule_params_.schedule_interval_).

This simplifies some of the tests: It is no longer necessary to
"prime the pump" by sending the first Observation before entering
the state of waiting for a scheduled send.

We also promote the constant kVeryLongTime from being in a test
to being part of the interface. With these two changes it is possible
to initialize a ShippingManager with the parameters
(kVeryLongTime, 0) and effectively get behavior that is the same
as not using a ShippingManager: namely that there are no scheduled sends--
sends only occur when you invoke RequestSendSoon() and in that case
they occur immediately. I want this behavior to be possible because
I want to use the ShippingManager in our TestApp.

Change-Id: Ib1fbec2f301976f4cceabb6f8cf9ed0baf27a32e