Adds the flag -repeat to the test_app.
This flag means to repeat the add-observations/send-to-shuffler
operation that many times. We use the flag in the end-to-end test.

Our goal is to allow the end-to-end test to test the case that
a single instance of ShippingManager is used more than once.
This would reproduce the bug that was fixed in
https://fuchsia-review.googlesource.com/#/c/cobalt/+/56445/.

Change-Id: Ie6295ca81a439c811b568f3337fe37ec7b2ce23f