Adds a  version of RequestSendSoon() that takes a callback.
We add a new version of ShippingManager::RequestSendSoon() that
accepts a callback that will be invoked whenever the requested
send has completed. The intended use of this is in tests of the
Fuchsia FIDL service. In that case we want to be able to
have a test that is a FIDL client be able to determine if
a send to the Shuffler succeeded.

We also fix one of the cases of ShippingManagerTest that turned
out to be a little flaky.

Change-Id: I23dc8dd996df90bfceebbec80e7dc7af1413c60b