Eliminate use of FRIEND_TEST macro.
Because we were using the FRIEND_TEST macro we were including
a file from the gtest library in non-test code. This is unnecessary
as the FRIEND_TEST macro is trivial and can be replaced by
its definition.

My motiviation for doing this is that the build is failing on the
commit queue for
https://fuchsia-review.googlesource.com/#/c/manifest/+/58276/

The logs show:./../cobalt/algorithms/rappor/rappor_config_validator.h:27:10: fatal error: 'third_party/googletest/googletest/include/gtest/gtest_prod.h' file not found

I don't really understand why that CQ run is failing because the
dependency on gtest is not new. But I've been meaning to clean this up
anyway so I'm doing it now to see if it fixes the CQ.

Change-Id: I2ce98e9f084adaaa49808ef4eec3040cbd3c580c