Update git submodule third_party/glog.
Move to cobalt branch. Note: not syncing to latest because doing
so caused the following pre-processor error that I didn't want
to deal with:

This is in utilities.cc. GetTID() function. The problem seems
to be that __i386__ is not defined and it should be. But it is
not important that we are synced to HEAD for glog so I am not
dealing with it.

  // On Linux and MacOSX, we try to use gettid().
  #ifndef __NR_gettid
    #ifdef OS_MACOSX
      #define __NR_gettid SYS_gettid
    #elif ! defined __i386__
      #error "Must define __NR_gettid for non-x86 platforms"
    #else
      #define __NR_gettid 224
    #endif
  #endif
  static bool lacks_gettid = false;
  if (!lacks_gettid) {
    pid_t tid = syscall(__NR_gettid);
    if (tid != -1) {
      return tid;
    }
    // Technically, this variable has to be volatile, but there is a small
    // performance penalty in accessing volatile variables and there should
    // not be any serious adverse effect if a thread does not immediately see
    // the value change to "true".
    lacks_gettid = true;
  }

CB-60

Change-Id: I0f6ca22cd02350fe51302fffd7b508f6b2de0093