Start building all third_party libraries
We change the method by which Cobalt links againt third_party
libraries that are written in C++. Previously our strategy
was:

- We download pre-built versions of .so files from Google
Cloud storage. We only have prebuilts for linux x86-64.
- We link against those at build time and we deploy the
.so files to the server at deploy time.

Now we change the strategy to:
- We incorporate the source code for all of the projects into
our third_party using Git submodules.
- We build the code along with our build.
- We link statically to the libraries so we do not deploy
with .so files.

One consequence of this work is that we are now using updated
versions of ProtBuf, gRPC, googletest, goolgeapis and gflags
relative to the pre-builts.

As part of this work we also cleaned up the CMakeLists files
with respect to how the specify linkage against the libraries
and also how they specify invoking protoc. We created a bunch
of CMake macros in the top-level CMake file to encapsulate this.

Most of the changes are to CMakeLists.txt files. There are
also a small number of changes having to do with the change
in deployment strategy. There are also changes to three
C++ files that became necessary because we are compiling
against a newer version of Protobuf.

CB-60

Change-Id: I37d3efc08f689d3b7de2e2c4497eb143b109765e