Enables TLS on the ReportMaster.
We enable the ReportMaster to serve gRPC using TLS.
At this time there is no plan to use this configuration in
production because in production instead we use Google Cloud
Endpoints and it terminates gRPC. We are doing this
for several reasons:

(1) It is useful to be able to enable TLS on ReportMaster when
running locally. Primarily this is useful for testing the
TLS flow on the client-side. I also find it useful in
general to help understand TLS.

(2) The code to do this was already mostly in place and there
were some TODOs so this just cleans up some loose ends.

(3) Its possible that in the future we may want to terminate
TLS in the ReportMaster or maybe we want to use TLS from
some intermediary like a load ballancer or Google Cloud
Endpoints.

Change-Id: Ia4d9ecbe75ec70e903cd01ff3997d1b0ea7bdd4d