Adds support for TLS when runnig Cobalt locally.
This includes when running the end-to-end test locally
and when running the  manual demo.

Also partially adds suppor for TLS when running Cobalt
on a devel cluster. (That part is not quite done.)

- There are two new files in the end_to_end test directory:
localhost.crt and localhost.key. The cert uses the CN
"localhost". If these are deployed on the server-side and
the client uses localhost.crt for the root certs and the
client connects to the server at "localhost" then TLS works
locally.

- When the flag --use_tls=true is passed to any of

./cobaltb.py start test_app
./cobaltb.py start report_client
./cobaltb.py test --tests=e2e

Then the clients will use tls and will, by default, use
localhost.crt for the root certs.

- When the flag --use_tls=true is passed to either of

./cobaltb.py start shuffler
./cobaltb.py start report_master
./cobaltb.py test --tests=e2e

then the server will use tls and will, by default, use
localhost.crt and localhst.key.

- We also add a new command cobaltb.py generate_cert.
This generates a self-signed cert that has the additional
feature of having an "Alternate Name" which is an IP address.
This is necessary in order to:
(a) Use tls when running locally but when the client connects
to the server using an IP address instead of "localhost". This
is useful in order to be able to run the server locally and
run the client on Fuchsia in QEMU.
(b) Use tls in the devel cluster. In this case the client connects
to the server using an IP address.

- The file self-signed-with-ip.cnf is an openssl configuration file
template that is used for the generate_cert command.

Change-Id: I1c407d5248e5c2498e468fd80d50e60a62037e92