Adds random_id field to Observation.
This is in support of the work to make the add-observation operation
idempotent. In this CL we add the field and populate it on the
client side. We do not actually make use of the field yet on
the server side. That will be in the next CL.

We have a couple of tests which explicitly use the size in bytes
of an Observation. Since we are now increasing Observation size
by nine bytes we needed to update those tests.

CB-48

Change-Id: I52e60de96d694700b0f079d3a2558419975b457f