Adds git submodule for "benchmark".
Benchmark is a library used by the gRPC build. We do not use it at runtime.
We could have removed all references to it from the gRPC CMake file but
it is less invasive to simply bring Benchmark into our build. It
is already a third_party repo in Fuchsia.

In fact the Cobalt CMake file already does assume that we have brought
benchmark in as a third_party library but I forgot to actually do that.
The repercussion is that we see a build warning when building gRPC but
this is not fatal and is easily ignored. So the real purpose of this
CL is to get rid of that warning.

Change-Id: I22f383e10ab129a44eea75568aeb538960359175