Start using the new random_id field from Observation.
We refactor the row keys used by the Observation store. We replace
the last two segments of the row key <time>:<random> with two
new segments <random>:<hash>. The new random is a 64-bit random number.
If the field random_id() is present in the Observation then that will
be used for this field, otherwise a new random number will be generated.
The new <hash> field is a 32-bit hash of the Observation.

CB-48

Change-Id: Iadbfe8b68fe176f0bf8b4f91b5acbca643d2ed48