Adds test of the idempotency of add-observation() to the e2e test.
We add the flag -num_adds_per_observation to test_app and we set
the value of this flag to 3 in the end-to-end test. This means
that during the end-to-end test each generated Observation will
be sent to the Shuffler 3 times. We alos add some VLOG(5) level
logging to manually check that the multiple adds and multiple
receives are actually occuring.

CB-48

Change-Id: Ib9cb405070a0c509c76be5846f54da24d0b0b766