Adds tooling to support using TLS on a devel cluster.
When running the Shuffler and the ReportMaster on GKE in a
development cluster (as opposed to a production cluster) we usually
want to use self-signed certificates. This CL adds some tooling
around creating and uploading self-signed certificates.

Note that most of the tooling already existed and was sufficient
for using TLS in other scenarios, namely in a production cluster
and when running Cobalt locally. Here we just needed to add a
few loose ends to support the devel cluster use case.

- We add the flags --shuffler_root_certs and
--report_master_root_certs to specify the location of the self-signed
cert to be used as the root CA file. Previously there was a single
flag --tls_root_certs but we needed separate ones for the
Shuffler and ReportMaster because they have different self-signed
certs since they have different host-names.

- We enchanced the command cobaltb.py generate_cert to generate
a self-signed certificate that put the host nome in a Subject
Alternative Name field. Previoulsy we were putting the host
name in the Common Name (CN) field and putting the IP address
in the Subject Alternative Name field. But that is not sufficient
because the C++ version of gRPC is strict about checking the
host name and the strict rules specify that if the cert contains
any Subject Alternative Names then the Common Name should not
be used.

Change-Id: Id97feda1f3a85f7c3c1e071f38aaabc35a200f87