String RAPPOR: Implements ExtractEstimatedBitCountRatios.
This function builds the column vector b for which the goal of the
RAPPOR LASSO analysis is to find an approximate solution to
Ax = b.

In terms of the R code this is the vector produced by converting
the matrix that is the output of the function EstimateBloomCounts()
in decode.R into a single column vector.

The code to produce the information that goes into forming this vector
already existed so in this CL all we are doing is forming the appropriate
ratios and coercing that data into a column vector.

We also include in the comments at the bottom of rappor_analyzer.cc
a mathematical explanation for the formula used to define the
estimated bit count ratio vector.

Change-Id: Ied18a1b1c8b20faa3048028d60113d010a070b2b