String RAPPOR: Adds rough first draft of code that uses lossmin to compute estimates.
This code is not intended to be used as-is for Cobalt production. My
expectation is that mironov@ will re-write this code corectly.

We use the ParallelBoostingWithMomentum minimizer from the lossmin library
with a LinearRegressionLossFunction to compute an estimate for x in the
equation Ax=b.

We were not successfuly at incorporating the l1 norm into the loss function.
When we attempted to do that the algorithm did not converge.

Even though this is not expected to be the final version of this code, this CL
serves several purposes:

- It incorporates the lossmin library into rappor_analyzer.cc
- It shows how to use the lossmin classes in this context.
- It includes some detailed notes about some of my experiences
with the library.
- It includes a framework for testing this code. I have added a very simple
test in rappor_analyzer_test.cc that I call ExperimentWithAnalyze(). It is
not so much a test as it is an experiment that invokes the Analyzer() method,
checks that the algorithm converges, checks that the result vector has
the right properties, and then prints out the true counts and the computed
estimates to the console. Here is an example of the output:

-------------------------------------
p=0, q=1, only candidate 5
True counts: 0 0 0 0 0 100 0 0 0 0
  Estimates: 0 0 0 0 0 100 0 0 0 0
-------------------------------------
p=0, q=1, several candidates
True counts: 0 20 0 0 20 0 0 0 0 60
  Estimates: 0 24 2 -3 16 -5 -5 -2 9 65
-------------------------------------
p=0.1, q=0.9, only candidate 5
True counts: 0 0 0 0 0 100 0 0 0 0
  Estimates: -4 5 5 -12 0 95 11 10 -5 0
-------------------------------------
p=0.1, q=0.9, several candidates
True counts: 0 20 0 0 20 0 0 0 0 60
  Estimates: 2 21 -2 -4 21 3 -2 1 2 55

Change-Id: I373182db5e2e4a2e0c00e269fab667cde0e83511