Fix build of config_parser.
I'm pretty sure that this flake:
https://ci.chromium.org/swarming/task/3908283c1c01ba10?server=chromium-swarm.appspot.com

is due to the fact that add_custom_command() for CONFIG_PARSER_BIN did
not have a DEPENDS on CONFIG_PB_GO_FILES. So I added that.

While I was at it I fixed a couple of other things:

(1) I renamed config.proto to cobalt_config.proto so that the file name
matches the message name.

(2) In the definition of CONFIG_PB_GO_FILES I fixed the spelling of
one of the files. This has been broken for a while--not sure why
we never noticed. It should be report_configs.pb.go not reports.pb.go.

(3) I removed the unnecessary set_source_files_properties() for
YAMLPB_TEST_PB_GO. Firstly, as I suggested in a code review, this
statement was not doing anything becase the variable
YAMPB_TEST_PB_GO was not set yet. At the time that
set_source_files_properties() executed the value of the variable
was the empty string. But secondly I realized that the statement
was not necessary any way. That statement is used when a file is
listed as a source file in an add_executable() but that is not
the case here.

(4) I removed DEPENDS ${CONFIG_PARSER_TEST_PB_GO} since the
varaible CONFIG_PARSER_TEST_PB_GO is never defined.

Change-Id: I73fb95cfea49471e82bccba1755229426c77fdd5