Adds SystemProfile.
- Adds the SystemProfile proto message. This will be collected by the
Encoder and added to each Envelope. The Shuffler will copy it into
each ObservationMetadata. The Analyzer will store it with each
Observation.

- Adds the SystemData class. Currently this is used only to populate
and give access to a SystemProfile. We plan for its use to be more
general: It will be the means by which the Encoder client queries all
information about the system. In addition to the SystemProfile it will
also provide stateful information such as the current memory usage.

Change-Id: I3d781b071a68eac6260f8aa5a09e161983b43b08