ObservationStore: Read and Write SystemProfile
- The ObservationStore now has a new column for storing a
serialized SystemProfile with each Observation, if there
is one when the Observation is written to the store.

- QueryObservations() has a new parameter to specify whether or not
to fetch this new column.

- QueryResult has been refactored: The day_index field has been
replaced by an ObservationMetadata field that includes the day_index,
and optionally the SystemProfile. In the future we may wish to
add additional data here.

- The observation hash component of the row keys has been
augmented to be a hash of the pair (Observation, ObservationMetadata)
since now there may be information contained in ObservationMetadata
that is not reflected in other parts of the row key.

- There is not yet any support for running reports that include
SystemProfile.

Change-Id: I7e06f7820268cfb0a3652ab9a3945506f8491beb