Changes implementation of the Ping() method.
The Ping method is a temporary hack to get around
https://github.com/google/google-api-cpp-client/issues/48

We just need to do some interaction with GCS other than an
upload. Previously I implemented Ping() as listing the
buckets in a project. Here I change the implementation of
Ping() to fetching the metadata for a bucket. This is better
since it requires only the name of a bucket and not the
name of a project.

We also removed the name of a real bucket that we were using
for testing and replaced it with a place-holder.

Change-Id: I9eb8e7d92d3946b966a367f276d9b1420d140d17