Expose extra out parameter in internal function.
We change the signature of the internal function StartReportNoAuth()
so that it now returns the internal ReportId. Previoulsy it
computed the internal ReportId but only used it to then compute
the external version of the report ID and did not return the
internal version. This extra parameter is not used from with
ReportMasterService. But we will use it in the next CL
from within the ReportScheduler.

Change-Id: Ib1439fde75a284b4423dcfabdf6ddf007069a42e